interaction-service/requirements.txt
Django==4.2.27
djangorestframework==3.15.2
pymongo==4.10.1
gunicorn==23.0.0
django-cors-headers==4.6.0
python-dotenv==1.0.1
